Some GAM code reorganization
